requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.2
django-appconf==1.0.3
django-imagekit==4.0.2
gunicorn==20.0.4
idna==2.8
pilkit==2.0
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sodapy==2.0.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
psycopg2-binary==2.7.7
